[SP-3631] move HttpMock to Helpers folder
